Bug 1800611 - part 4: Rename `copy` module in order not to override the system one
