+++ maven_lambda/copy_to_bucket.py
