+++ maven_lambda/test/test_copy_to_bucket.py
@@ -1,7 +1,7 @@
 from unittest.mock import MagicMock
 import pytest
 
-from maven_lambda.copy import (
+from maven_lambda.copy_to_bucket import (
     lambda_handler,
     s3_object_has_more_than_one_version,
     NotFound,
@@ -26,7 +26,7 @@ def test_lambda_handler(monkeypatch):
     monkeypatch.setattr('boto3.client', lambda _: s3_mock)
     monkeypatch.setattr('os.environ', {'TARGET_BUCKET': 'some_target_bucket'})
 
-    monkeypatch.setattr('maven_lambda.copy.s3_object_has_more_than_one_version', lambda _, __, ___: False)
+    monkeypatch.setattr('maven_lambda.copy_to_bucket.s3_object_has_more_than_one_version', lambda _, __, ___: False)
     assert lambda_handler(event, context) == {"statusCode": 200}
     s3_mock.copy_object.assert_called_once_with(
         Bucket='some_target_bucket',
@@ -38,13 +38,13 @@ def test_lambda_handler(monkeypatch):
     )
     s3_mock.reset_mock()
 
-    monkeypatch.setattr('maven_lambda.copy.s3_object_has_more_than_one_version', lambda _, __, ___: True)
+    monkeypatch.setattr('maven_lambda.copy_to_bucket.s3_object_has_more_than_one_version', lambda _, __, ___: True)
     assert lambda_handler(event, context) == {"statusCode": 409}
     s3_mock.copy_object.assert_not_called()
 
     def fail(_, __, ___):
         raise NotFound()
-    monkeypatch.setattr('maven_lambda.copy.s3_object_has_more_than_one_version', fail)
+    monkeypatch.setattr('maven_lambda.copy_to_bucket.s3_object_has_more_than_one_version', fail)
     assert lambda_handler(event, context) == {"statusCode": 404}
     s3_mock.copy_object.assert_not_called()
